Replaced the preferences system

+++ src/core/util.c
@@ -206,7 +206,7 @@ float lerp_angle(float alpha, float beta, float t)
     v2d_t a = v2d_new(cosf(alpha), sinf(alpha));
     v2d_t b = v2d_new(cosf(beta), sinf(beta));
     v2d_t c = v2d_lerp(a, b, t);
-    return atan2(c.y, c.x);
+    return atan2f(c.y, c.x);
 }
